net/i40e: fix clang build error with 16B descriptors
When compiling with 16B descriptor support enabled, clang compiles gave an
error, complaining that the final parameter of _mm256_blend_epi32() had to
be an immediate value (i.e. compile-time constant). While it appears that
GCC was able to convert the constant variable value "fdir_blend_mask" into
the blend call, clang was not doing so. To guarantee the use of an
immediate we convert the variable value to a #define.

Fixes: 7d087a0a8b8e ("net/i40e: support flow director on AVX Rx")
Cc: harry.van.haaren@intel.com

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>