net/ice: fix FDIR flow type conflict
Flow type "IPv4 + UDP" or "IPv4 + TCP" is conflict with "IPv4 + any"
flow type. If a rule for IPv4 + any is created, we should reject any rule
for IPv4 + UDP otherwise the first rule may be impacted, same decision
should be made on a reverse order.
For IPv6 and IPv4 GTPU inner case, we have the same limitation.

Fixes: 109e8e06249e ("net/ice: configure HW flow director rule")

Signed-off-by: Qi Zhang <qi.z.zhang@intel.com>
Tested-by: Beilei Xing <beilei.xing@intel.com>
Acked-by: Beilei Xing <beilei.xing@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>