doc: drop EAL command-line reference from FreeBSD GSG
The EAL command-line reference section duplicated material covered in a
previous chapter, was out of date, and generally is unnecessary as options
are best queried directly by passing "--help" to a DPDK binary. Therefore
we can drop this section to avoid having to try and keep it up-to-date in
the future.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>