build: clean up experimental syms check
This patch cleans up the meson build files in lib and drivers by moving the
custom target for checking the experimental syms next to the abi compat
checks. This also removes the dependency on the check for the shared build,
which was not required by anything, but was only added to force the
experimental syms check run.

Signed-off-by: Kevin Laatz <kevin.laatz@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>