net/i40e/base: revert Fix missing "fall through" comments
These fall throughs are not needed by the compiler and
are more distracting.  The fall through comment is
needed when there is a block of code in a switch statement
and then falls through to the next case.  It is not
needed if it's just a case and automatically falls through
on the next line.

By reverting this patch we also line up the file with the
upstream version.

Signed-off-by: Alice Michael <alice.michael@intel.com>
Reviewed-by: Kirsher Jeffrey T <jeffrey.t.kirsher@intel.com>
Signed-off-by: Xiaolong Ye <xiaolong.ye@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>