net/i40e/base: backport style changes from upstream Linux
The upstream virtchnl.h has a few minor style differences to what is
out-of-tree, and there is no real reason we should prefer the current
style over whats upstream.

Fix the VIRTCHNL_CHECK_STRUCT_LEN macro whitespace, and fix the comment
about "this macro is" to be "these macros are".

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Reviewed-by: Kirsher Jeffrey T <jeffrey.t.kirsher@intel.com>
Signed-off-by: Xiaolong Ye <xiaolong.ye@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>