net/i40e/base: changeed code wrapping for CARLSVILLE_HW
As pointed out by Arkadiusz, the code is not wrapped properly and if
CARLSVILLE_HW is not defined the code will not compile.  In addition,
noticed that the we were taking the if condition majority of the time so
flip the condition to only test for the device that requires the
workaround when updating stats.

Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Reviewed-by: Paul M Stillwell Jr <paul.m.stillwell.jr@intel.com>
Reviewed-by: Aleksandr Loktionov <aleksandr.loktionov@intel.com>
Reviewed-by: Michael Alice <alice.michael@intel.com>
Signed-off-by: Xiaolong Ye <xiaolong.ye@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>