net/i40e/base: change long redundant define names
There is no need to create define names that are practically a full
sentence.  If you want to describe in detail what the define is for, you
a code comment.  Define name can be abbreviated and should not contain
verbs or "actions", so abbreviate as much as possible.

These defines that were abbreviated could potentially get even shorter,
but not without adequate code comments.

Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Reviewed-by: Aleksandr Loktionov <aleksandr.loktionov@intel.com>
Reviewed-by: Galazka Krzysztof <krzysztof.galazka@intel.com>
Reviewed-by: Michael Alice <alice.michael@intel.com>
Signed-off-by: Xiaolong Ye <xiaolong.ye@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>