drivers: add a HW quirk for register definitions
V1000/R1000 processors are using the same PCI ids for the network
device as SNOWYOWL processor but has altered register definitions
for determining the window settings for the indirect PCS access.
Add support to check for this hardware and if found use the new
register values.

Added a new routine rte_pci_search_device to pci driver to search
for a device.

Signed-off-by: Selwin Sebastian <Selwin.Sebastian@amd.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>