examples/l3fwd: improve readability for destination lookup
The functions to lookup ipv4 and ipv6 were both using opaque
pointers (void *) when they should use a typed pointer instead.
The ip headers are not modified during lookup.

Get rid of unnecessary cast on the return from the function.
Replace complex trigraph expression with simple if to improve
readability.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>