net/mlx5: fix vxlan-gpe item translation
Currently, when using vxlan-gpe or vxlan item in the flow
both are being treated the same with flags 0x8 in vxlan
header. which mean the matching of the item vxlan-gpe
will match any vxlan packet.

This fix the translation of vxlan gpe item into pmd flow
item. which will by default set the flags to vxlan-gpe
to be 0xc.

Fixes: 3d694341 ("net/mlx5: add Direct Verbs validation function")
Cc: stable@dpdk.org

Signed-off-by: Raslan Darawsheh <rasland@mellanox.com>
Acked-by: Matan Azrad matan@mellanox.com
Signed-off-by: 0-day Robot <robot@bytheb.org>