net/octeontx2: update link information for loopback port
loopback devices are exposed as ethdev device in octeontx2.
dpdk link update APIs updating the link information for cgx
ports but skipping loopback ports.

When stack uses loopback port for forwarding, packets are
dropped as link status is down. Link information need to be
updated for loopback ports to avoid it.

Fixes: 38f566280a ("net/octeontx2: add link stats operations")
Cc: stable@dpdk.org

Signed-off-by: Ashish Gupta <ashishg@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>