net/mlx5: fix the hairpin queue capacity
The hairpin TX/RX queue depth and packet size is fixed in the past.
When the firmware has some fix or improvement, the PMD will not
make full use of it. And also, 32 packets for a single queue will not
guarantee a good performance for hairpin flows.
The parameter of hairpin queue setup needs to be adjusted. Number of
packets of a single queue should be the maximum supported value, and
the maximum single packet size should support the standard jumbo
frame with 9KB. In the meanwhile, there is no need to support the
max capacity of a single packet because the memory consumption should
also be taken into consideration.

Fixes: e79c9be91515 ("net/mlx5: support Rx hairpin queues")
Cc: orika@mellanox.com
Cc: stable@dpdk.org

Signed-off-by: Bing Zhao <bingz@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>